add -Wall, -Wextra, and fix all warnings which don't indicate an issue

+++ Realm/RLMObject.mm
@@ -63,7 +63,7 @@ NSString *const RLMPropertyAttributeRequired = @"RLMPropertyAttributeRequired";
         }
         // FIXME - more validation for each property type
         
-        for (int i = 0; i < array.count; i++) {
+        for (NSUInteger i = 0; i < array.count; i++) {
             [obj setValue:array[i] forKeyPath:[properties[i] name]];
         }
     }
